Connect project to database

+++ BookingService.Housing.Models/Stay.cs
@@ -5,7 +5,7 @@ public sealed class Stay
     public int Id { get; set; }
     public DateOnly From { get; init; }
     public DateOnly To { get; init; }
-    public decimal TotalPrice { get; init; }
+    public double TotalPrice { get; init; }
     public StayStatus Status { get; init; }
     
     public int UnitId { get; init; }
