+++ BookingService.Shared/Filters/HousingFilterOptions.cs
@@ -8,6 +8,6 @@ public sealed class HousingFilterOptions
     public string? Name { get; init; }
     public string? City { get; init; }
     public string? Country { get; init; }
-    public decimal? MaxPrice { get; init; }
-    public decimal? MinPrice { get; init; }
+    public double? MaxPrice { get; init; }
+    public double? MinPrice { get; init; }
 }
